pyproject.toml
[project]
name = "BellaVista-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Wecreds", email = "Wecredsbr@gmail.com"},
]
dependencies = [
    "django>=5.1.2",
    "dotenv>=0.0.5",
    "django-cors-headers>=4.5.0",
    "python-dotenv>=1.0.1",
    "djangorestframework>=3.15.2",
    "Pillow>=11.0.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-spectacular>=0.27.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[tool.pdm.scripts]
dev = "python manage.py runserver 0.0.0.0:19003"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
shell = "python manage.py shell"
shellp= "python manage.py shell_plus"
populate = "pdm run python manage.py populate"
clear.composite = [
    'find . -name "__pycache__" -type d -exec rm -r {} +',
    'find . -path "*/migrations/*.pyc" -delete',
    'find . -path "*/migrations/*.py" -not -name "__init__.py" -delete',
    'rm -rf __pypackages__ pdm.lock',
    'rm db.sqlite3',
]